철자수정, thread 수정 적용
