+++ Refrigerator/app/src/main/java/refrigerator/FileManage.java
@@ -1,4 +1,4 @@
-package refirgerator;
+package refrigerator;
 
 import java.io.*;
 import java.util.*;
