+++ Refrigerator/app/src/main/java/refrigerator/Food.java
@@ -1,4 +1,4 @@
-package refirgerator;
+package refrigerator;
 
 import java.util.Objects;
