+++ Refrigerator/app/src/main/java/refrigerator/Main.java
@@ -1,11 +1,11 @@
-package refirgerator;
+package refrigerator;
 
 import java.util.*;
 
 public class Main {
     //public static void main(String[] args)
     {
-        // 여기말고 메인프레임 메인에 작성
+        // 여기말고 메인트레임 메인에 작성
 
     }
 }
