+++ Refrigerator/app/src/main/java/refrigerator/Refrigerator.java
@@ -1,4 +1,4 @@
-package refirgerator;
+package refrigerator;
 import java.util.*;
 
 public class Refrigerator {
